Fix "DeviceId" Casting
DeviceId is converted from Int to Long

+++ CoronavirusHerdImmunity/app/src/main/java/com/example/coronavirusherdimmunity/MainActivity.java
@@ -68,7 +68,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
                         bolts.Task.callInBackground(new Callable<Object>() {
                             @Override
                             public Object call() throws Exception {
-                                int deviceId = new PreferenceManager(mContext).getDeviceId();
+                                Long deviceId = new PreferenceManager(mContext).getDeviceId();
                                 ApiManager.registerPushToken(deviceId, token, new PreferenceManager(mContext).getAuthToken());
                                 return null;
                             }
@@ -228,7 +228,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
     private void writeQRCode() {
         ImageView qrImage = (ImageView) findViewById(id.qr_code);
 
-        int deviceId = new PreferenceManager(mContext.getApplicationContext()).getDeviceId();
+        Long deviceId = new PreferenceManager(mContext.getApplicationContext()).getDeviceId();
 
         QRCodeGenerator generator = new QRCodeGenerator(mContext);
         generator.generateQRCode(deviceId, qrImage);
